+++ Rdmp.Core/Icons/IconOverlays/Overlays.Designer.cs
@@ -265,5 +265,16 @@ namespace Rdmp.Core.Icons.IconOverlays {
                 return ((System.Drawing.Bitmap)(obj));
             }
         }
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
+        internal static System.Drawing.Bitmap BigE
+        {
+            get
+            {
+                object obj = ResourceManager.GetObject("BigE", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
     }
 }