+++ Rdmp.Core/Icons/IconProvision/CatalogueIconProvider.cs
@@ -35,8 +35,10 @@ namespace Rdmp.Core.Icons.IconProvision
         protected List<IObjectStateBasedIconProvider> StateBasedIconProviders = new List<IObjectStateBasedIconProvider>();
 
         protected readonly EnumImageCollection<RDMPConcept> ImagesCollection;
-
+        protected readonly CatalogueStateBasedIconProvider CatalogueStateBasedIconProvider;
         private DatabaseTypeIconProvider _databaseTypeIconProvider = new DatabaseTypeIconProvider();
+
+        public Bitmap ImageUnknown => ImagesCollection[RDMPConcept.NoIconAvailable];
 
         public CatalogueIconProvider(IRDMPPlatformRepositoryServiceLocator repositoryLocator,
             IIconProvider[] pluginIconProviders)
@@ -45,7 +47,7 @@ namespace Rdmp.Core.Icons.IconProvision
             OverlayProvider = new IconOverlayProvider();
             ImagesCollection = new EnumImageCollection<RDMPConcept>(CatalogueIcons.ResourceManager);
 
-            StateBasedIconProviders.Add(new CatalogueStateBasedIconProvider(repositoryLocator.DataExportRepository, OverlayProvider));
+            StateBasedIconProviders.Add(CatalogueStateBasedIconProvider = new CatalogueStateBasedIconProvider(repositoryLocator.DataExportRepository, OverlayProvider));
             StateBasedIconProviders.Add(new ExtractionInformationStateBasedIconProvider());
             StateBasedIconProviders.Add(new ExtractableColumnStateBasedIconProvider(OverlayProvider));
             StateBasedIconProviders.Add(new CheckResultStateBasedIconProvider());
@@ -201,7 +203,7 @@ namespace Rdmp.Core.Icons.IconProvision
             }
 
 
-            return ImagesCollection[RDMPConcept.NoIconAvailable];
+            return ImageUnknown;
 
         }